Add routing and links to the signup page
This required a signficant reworking of the
existing components due to the way the various
parts of react-router-dom work.

For example the call to useLocation cannot sit
in the same component that contains Router. This
meant I had to create a 'Main' component to use
useLocation, with the Router being in the app
itself.