Implement signup logic
Signing up now sends a properly formed post request
to the server.

Based on the format of the response specified by the
server API, I decided that the best way of handling
things if the sign up is successful is to require
to log in manually rather than automatically doing
it for them. This makes things a little easier for
me.