Add test for header showing welcome message on login
At present, this test does not work. I cannot work out how to
mock/use/whatever the context with custom values without having
to manipulate the functions within the context directly. Given
this is supposed to be a unit test, I really don't want to have to
do that.

At the moment, I'm leaning towards it being an issue with the way
I've implemented the context itself, but I can't see another way
of doing it while still allowing me to pass functions as part of
the context state.

I will continue investigating, but for the moment the offending
test is disabled.